Make StreamingSheet#iterator() reusable
StreamingSheet#iterator() returns an Iterator that starts from the beginning of the Sheet on each call
