+++ src/main/java/com/github/pjfanning/xlsx/impl/ooxml/HyperlinkData.java
@@ -1,5 +1,7 @@
 package com.github.pjfanning.xlsx.impl.ooxml;
 
+import java.util.Objects;
+
 public class HyperlinkData {
 
   private final String id;
@@ -34,5 +36,27 @@ public class HyperlinkData {
 
   public String getTooltip() {
     return tooltip;
+  }
+
+  @Override
+  public boolean equals(Object o) {
+    if (this == o) return true;
+    if (!(o instanceof HyperlinkData)) return false;
+    HyperlinkData that = (HyperlinkData) o;
+    return Objects.equals(id, that.id)
+        && Objects.equals(ref, that.ref)
+        && Objects.equals(location, that.location)
+        && Objects.equals(display, that.display)
+        && Objects.equals(tooltip, that.tooltip);
+  }
+
+  @Override
+  public int hashCode() {
+    return Objects.hash(
+        id,
+        ref,
+        location,
+        display,
+        tooltip);
   }
 }
